Refactor project contract loaders
