+++ src/hooks/v2v3/contractReader/ProjectPrimaryEthTerminalAddress.ts
@@ -2,7 +2,7 @@ import { ETH_TOKEN_ADDRESS } from 'constants/v2v3/juiceboxTokens'
 import { V2V3ContractName } from 'models/v2v3/contracts'
 import useV2ContractReader from './V2ContractReader'
 
-export function useProjectPrimaryEthTerminal({
+export function useProjectPrimaryEthTerminalAddress({
   projectId,
 }: {
   projectId: number | undefined
